requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
Django==3.1
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
djangorestframework==3.11.1
importlib-metadata==1.7.0
pillow==11.1.0
python-decouple==3.3
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.1.0
